+++ src/store/dispatcher.mjs
@@ -33,20 +33,22 @@ export function createStore( reducer, initial_state ) {
 
   function dispatch( action ) {
     let local_dispatch_id = ++dispatch_id
+    let is_unchanged = false
 
     try {
-      // @todo put profiling behind debug flag
       store.is_dispatching = true
       debug( `dispatch ${ action.type } start` )
       const new_state = reducer( current_state, action )
       debug( `dispatch ${ action.type } finish` )
-      if( validateState( new_state ) ) {
-        debug( 'action', action )
+      debug( 'action', action )
+      if( current_state === new_state ) {
+        debug( 'state is unchanged' )
+        is_unchanged = true
+      } else if( validateState( new_state ) ) {
         debug( 'state', new_state )
       } else {
         console.error( 'Validator failed on new state' )
-        console.info( 'current_state',current_state )
-        console.info( 'action', action )
+        console.info( 'current_state', current_state )
         console.info( 'new_state', new_state )
         console.info( 'errors', validateState.errors )
       }
@@ -55,12 +57,16 @@ export function createStore( reducer, initial_state ) {
       store.is_dispatching = false
     }
 
-    for( let listener of listeners.slice() ) {
-      if( dispatch_id !== local_dispatch_id ) {
-        console.warn('parallel dispatch, early abort')
-        break
+    if( ! is_unchanged ) {
+      debug( 'listeners start' )
+      for( let listener of listeners.slice() ) {
+        if( dispatch_id !== local_dispatch_id ) {
+          console.warn('parallel dispatch, early abort')
+          break
+        }
+        listener()
       }
-      listener()
+      debug( 'listeners finish' )
     }
     return action
   }